Fix file existence check

+++ install.sh
@@ -1,11 +1,10 @@
 #!/bin/bash
 
 git submodule update --init --recursive
-source activate py2Env
-pip install --requirement python-package-requirement.txt
 
 #Download Wikipedia dump
-mkdir ./data/wikipedia/dump/en -p
-wget  -O ./data/wikipedia/dump/en/dump.xml.bz2 https://dumps.wikimedia.org/enwiki/latest/enwiki-latest-pages-articles1.xml-p10p30302.bz2
+if [ ! -e "./data/wikipedia/dump/en/dump.xml.bz2" ]; then
+    mkdir ./data/wikipedia/dump/en -p
+    wget  -O ./data/wikipedia/dump/en/dump.xml.bz2 https://dumps.wikimedia.org/enwiki/latest/enwiki-latest-pages-articles1.xml-p10p30302.bz2
+fi
 
-mkdir ./snorkel/sentimantic/data/samples/ -p
